+++ app/(base)/new/EventsFromImage.tsx
@@ -4,34 +4,11 @@ import EventsError from "./EventsError";
 import { AddToCalendarCard } from "@/components/AddToCalendarCard";
 // import { generatedIcsArrayToEvents } from "@/lib/icalUtils";
 import { type AddToCalendarButtonProps } from "@/types";
-import { getPrompt, getSystemMessage } from "@/lib/prompts";
-
-// parse the response text into array of events. response format is:
-interface Response {
-  events: Event[]; // An array of events.
-}
-
-interface Event {
-  name: string; // The event's name.
-  description: string; // Short description of the event, its significance, and what attendees can expect.
-  startDate: string; // Start date in YYYY-MM-DD format.
-  startTime?: string; // Start time, if applicable (omit for all-day events).
-  endDate: string; // End date in YYYY-MM-DD format.
-  endTime?: string; // End time, if applicable (omit for all-day events).
-  timeZone: string; // Timezone in IANA format.
-  location: string; // Location of the event.
-}
-
-export const extractJsonFromResponse = (response: string) => {
-  // OpenAI returns a JSON code block starting with ```json and ending with ```
-  const start = response.indexOf("```json");
-  const end = response.lastIndexOf("```");
-  if (start === -1 || end === -1) {
-    return null;
-  }
-  const jsonString = response.slice(start + 7, end);
-  return JSON.parse(jsonString) as Response;
-};
+import {
+  addCommonAddToCalendarPropsFromResponse,
+  getPrompt,
+  getSystemMessage,
+} from "@/lib/prompts";
 
 const blankEvent = {
   options: [
@@ -122,50 +99,9 @@ export default async function EventsFromImage({
     return <EventsError rawText={imageUrl} />;
   }
 
-  let events = [] as AddToCalendarButtonProps[];
+  const events = addCommonAddToCalendarPropsFromResponse(response);
 
-  try {
-    const res = extractJsonFromResponse(response);
-    if (!res) {
-      throw new Error("Failed to parse response");
-    }
-    events = res.events.map((event) => {
-      return {
-        options: [
-          "Apple",
-          "Google",
-          "iCal",
-          "Microsoft365",
-          "MicrosoftTeams",
-          "Outlook.com",
-          "Yahoo",
-        ] as
-          | (
-              | "Apple"
-              | "Google"
-              | "iCal"
-              | "Microsoft365"
-              | "MicrosoftTeams"
-              | "Outlook.com"
-              | "Yahoo"
-            )[]
-          | undefined,
-        buttonStyle: "text" as const,
-        name: event.name,
-        description: event.description,
-        location: event.location,
-        startDate: event.startDate,
-        endDate: event.endDate,
-        startTime: event.startTime || undefined,
-        endTime: event.endTime || undefined,
-        timeZone: event.timeZone,
-      };
-    });
-  } catch (e: unknown) {
-    console.log(e);
-  }
-
-  if (events.length === 0) {
+  if (!events || events.length === 0) {
     return (
       <>
         <EventsError rawText={imageUrl} response={response || undefined} />
